Data.SOAS

Documentation

data AnnSig ann (sig :: Type -> Type -> Type) scopedTerm term

Constructors

AnnSig (sig scopedTerm term) ann

data AnnBinder ann (binder :: S -> S -> Type) (n :: S) (l :: S)

Constructors

AnnBinder (binder n l) ann

class TypedBinder (binder :: S -> S -> Type) t where

Methods

addBinderTypes :: forall (n :: S) (l :: S). binder n l -> t -> NameMap n t -> NameMap l t

data MetaAppSig metavar scope term

Second-order signature for metavariable applications. This way, metavariables can be added to any other signature using Sum.

Constructors

MetaAppSig metavar [term]
A metavariable is always fully applied to a list of arguments.

pattern Node' :: forall sig (binder :: S -> S -> Type) metavar n t. sig (TypedScopedSOAS binder metavar sig n t) (TypedSOAS binder metavar sig n t) -> t -> TypedSOAS binder metavar sig n t

type TypedSOAS (binder :: S -> S -> Type) metavar (sig :: Type -> Type -> Type) (n :: S) t = AST (AnnBinder t binder) (AnnSig t (Sum sig (MetaAppSig metavar))) n

type TypedScopedSOAS (binder :: S -> S -> Type) metavar (sig :: Type -> Type -> Type) (n :: S) t = ScopedAST (AnnBinder t binder) (AnnSig t (Sum sig (MetaAppSig metavar))) n

type SOAS (binder :: S -> S -> Type) metavar (sig :: Type -> Type -> Type) (n :: S) = TypedSOAS binder metavar sig n ()

A second-order abstract syntax (SOAS) is generated from a signature by adding parametrised metavariables.

Note that here we also parametrise SOAS by the type of binders (patterns) binder.

type ScopedSOAS (binder :: S -> S -> Type) metavar (sig :: Type -> Type -> Type) (n :: S) = TypedScopedSOAS binder metavar sig n ()

A scoped version of SOAS. ScopedSOAS is to SOAS what ScopedAST is to AST.

newtype MetaSubst (binder :: S -> S -> Type) (sig :: Type -> Type -> Type) metavar t

A metavariable substitution is a pair of a metavariable name and its body.

Constructors

MetaSubst

Fields

metaSubst :: (metavar, MetaAbs binder sig t)

newtype MetaSubsts (binder :: S -> S -> Type) (sig :: Type -> Type -> Type) metavar t

A collection of metavariable substitutions (for simultaneous substitution | of multiple metavariables).

Constructors

MetaSubsts

Fields

metaSubsts :: [MetaSubst binder sig metavar t]

pattern MetaApp :: forall metavar (binder :: S -> S -> Type) (sig :: Type -> Type -> Type) n t. metavar -> [TypedSOAS binder metavar sig n t] -> t -> TypedSOAS binder metavar sig n t

A convenient pattern synonym for parametrised metavariables.

data MetaAbs (binder :: S -> S -> Type) (sig :: Type -> Type -> Type) t where

A body of a metavariable substitution for one metavariable.

Constructors

MetaAbs

Fields

:: forall (n :: S) (binder :: S -> S -> Type) (sig :: Type -> Type -> Type) t. NameBinderList 'VoidS n

A list of binders corresponding to metavariable arguments.

-> AST binder sig n

Term to substitute the metavariable with.

-> MetaAbs binder sig t

match

Arguments

:: forall (sig :: Type -> Type -> Type) (n :: S) (binder :: S -> S -> Type) (ext :: Type -> Type -> Type) t metavar. (Bitraversable sig, ZipMatchK sig, Distinct n, UnifiablePattern binder, SinkableK binder, Bitraversable ext, ZipMatchK (Sum sig ext), Bitraversable (AnnSig t sig), ZipMatchK (Sum (AnnSig t sig) ext), ZipMatchK (AnnSig t sig), Eq t, Ord metavar, TypedBinder binder t, ZipMatchK (Sum sig (MetaAppSig metavar)), ZipMatchK t)

=> Scope n
The current scope.

-> Map metavar ([t], t)

-> NameMap n t

-> TypedSOAS binder metavar sig n t
The left-hand side (with metavariables that we solve for).

-> AST (AnnBinder t binder) (AnnSig t (Sum sig ext)) n
The right hand side (rigid)

-> [MetaSubsts (AnnBinder t binder) (AnnSig t (Sum sig ext)) metavar t]

Match left-hand side (with metavariables) against the rigid right-hand side.

If matching is successful, it produces metavariable substitutions that when applied to LHS make it syntactically equal to RHS. For example, matching M[f x, g] = g (f x) produces substitution M[z₁, z₂] ↦ z₂ z₁

There may be more than one solution for matching, e.g. M[f x, f x] = f x can be solved with two different substitutions: 1. M[z₁, z₂] ↦ z₁ 2. M[z₁, z₂] ↦ z₂

Hence, this function produces a list of possible substitutions.

Inputs must be well-scoped, with complete variable contexts and correct operator annotations. Matching checks the types of the compared terms and the declared arities and types of metavariables. It does not implement an object-language type checker.

push :: forall (i :: S) (l :: S) (n :: S). NameBinder i l -> NameBinderList n i -> NameBinderList n l

O(n). Push a name binder into the end of a name binder list.

Should be in Control.Monad.Foil.

toNameMap :: forall (m :: S) a (l :: S). NameMap m a -> NameBinderList m l -> [a] -> NameMap l a

renameNameMap :: forall (n :: S) (l :: S) a. (Name n -> Name l) -> NameMap n a -> NameMap l a

Transport a name map along an injective renaming of its scope.

applyMetaSubsts

Arguments

:: forall (sig :: Type -> Type -> Type) metavar metavar' (n :: S) (binder :: S -> S -> Type) t. (Bifunctor sig, Eq metavar, Distinct n, CoSinkable binder, SinkableK binder, metavar ~ metavar')

=> Scope n
Scope of terms.

-> MetaSubsts (AnnBinder t binder) (AnnSig t (Sum sig (MetaAppSig metavar'))) metavar t
Metavariable substitutions.

-> TypedSOAS binder metavar sig n t
The original SOAS term.

-> TypedSOAS binder metavar' sig n t

Apply metavariable substitutions to a SOAS term.

Utils

concatNameBinderLists :: forall (i :: S) (l :: S) (n :: S). NameBinderList i l -> NameBinderList n i -> NameBinderList n l

O(n). Concatenate name binder lists.

Should be in Control.Monad.Foil.

withFreshNameBinderList :: forall (n :: S) a (i :: S) r. Distinct n => [a] -> Scope n -> NameBinderList i n -> NameMap n a -> (forall (l :: S). Distinct l => Scope l -> NameBinderList i l -> NameMap l a -> r) -> r

Generate fresh name binders for a list of things. This is useful to generate proper name binders of metavariable parameters.