Language.Lambda.Framework

Description

Framework for testing and validating pattern matching and unification.

Types

Parsing

parseMetavarBinders :: [String] -> Either String MetavarBinders

Parse a list of metavariable binder strings

matchUnificationConstraint :: [String] -> String -> Either String [MetaSubsts']

Match a unification constraint with given metavariable binders

isSolvedUnificationConstraint :: [String] -> String -> [String] -> Either String Bool

Check if a unification constraint is solved by specific substitutions

parseMetaSubst :: MetavarBinders -> String -> Either String MetaSubst'

Validation

validateProblem :: CanonicalProblem -> Either String ([(CanonicalSolution, [CanonicalConstraint])], [CanonicalSolution])

validateSolution :: [CanonicalConstraint] -> CanonicalSolution -> Either (CanonicalSolution, [CanonicalConstraint]) CanonicalSolution

solveUnificationConstraint :: MetaSubsts' -> CanonicalConstraint -> CanonicalConstraint

Apply substitutions to both sides of a unification constraint

isSolved :: CanonicalConstraint -> Bool

Check if a constraint is solved (left-hand side equals right-hand side)

TOML configuration

parseConfigAndValidate :: IO ()

printInvalidSolutionsWithConstraint :: (Foldable t, Show a) => (CanonicalSolution, t a) -> IO ()

Instances

class IsCanonicalMetavarBinders b where

Methods

toCanonicalMetavarBinders :: b -> Result MetavarBinders

fromCanonicalMetavarBinders :: MetavarBinders -> Result b

class IsCanonicalConstraint c where

Methods

toCanonicalConstraint :: MetavarBinders -> c -> Result CanonicalConstraint

fromCanonicalConstraint :: MetavarBinders -> CanonicalConstraint -> Result c

class IsCanonicalSubstitutions s where

Methods

toCanonicalSubstitution :: MetavarBinders -> s -> Result MetaSubsts'

fromCanonicalSubstitution :: MetavarBinders -> MetaSubsts' -> Result s

data SolutionComparison

Constructors

Incomparable

RightMoreGeneral

LeftMoreGeneral

Equivalent

solveAndCompareToReferenceSolutionsWith

Arguments

:: (IsCanonicalConstraint c, IsCanonicalSubstitutions s, IsCanonicalMetavarBinders b)

=> CanonicalProblem
The problem definition with constraints to solve

-> (b -> [c] -> Result [Solution b s])
User's algorithm function

-> Result [(CanonicalSolution, SolutionComparison)]

Solve a unification problem and compare the results to reference solutions using a user-implemented algorithm. The function takes a problem definition and a solving function, and returns a list of comparisons between the solutions found by the algorithm and the reference solutions provided in the problem definition.

Main entry point

main :: IO ()

Main function for testing and debugging the framework

solveByMatchingAndCompareToReferenceSolutionsWith :: CanonicalProblem -> Result [(CanonicalSolution, SolutionComparison)]