Language.Lambda.Impl

Description

Free foil implementation of the (lambda)-calculus (with pairs).

Free foil provides general definitions or implementations for the following:

Freely generated (from a simple signature) scope-safe AST.

Correct capture-avoiding substitution (see substitute).

Correct α-equivalence checks (see alphaEquiv and alphaEquivRefreshed) as well as α-normalization (see refreshAST).

Conversion helpers (see convertToAST and convertFromAST).

The following is generated using Template Haskell:

Convenient pattern synonyms.

ZipMatch instances (enabling general α-equivalence).

Conversion between scope-safe and raw term representation.

The following is implemented manually in this module:

Computation of weak head normal form (WHNF), see whnf.

Entry point, gluing everything together. See defaultMain.

Note: free foil does not (easily) support patterns at the moment, so only wildcard patterns and variable patterns are handled in this implementation.

Types

type Term = AST FoilPattern TermSig

Scope-safe λ-term representation in scope n.

type MetaTerm metavar (n :: S) t = TypedSOAS FoilPattern metavar TermSig n t

type MetaSubst' = MetaSubst (AnnBinder Type FoilPattern) (AnnSig Type (Sum TermSig (MetaAppSig MetavarIdent))) MetavarIdent Type

type MetaSubsts' = MetaSubsts (AnnBinder Type FoilPattern) (AnnSig Type (Sum TermSig (MetaAppSig MetavarIdent))) MetavarIdent Type

type MetavarBinder = (MetavarIdent, MetavarType)

type MetavarBinders = Map MetavarIdent MetavarType

data UnificationConstraint where

Constructors

UnificationConstraint :: forall (n :: S). Distinct n => Scope n -> NameBinderList 'VoidS n -> NameMap n Type -> (MetaTerm MetavarIdent n Type, Type) -> (MetaTerm MetavarIdent n Type, Type) -> UnificationConstraint

data TermSig scope term

Generated with mkSignature. A signature bifunctor, specifying the nodes of a syntax tree corresponding to Term.

Constructors

LamSig Type scope
Corresponds to Lam.

AppSig term term
Corresponds to App.

MetavarSig MetavarIdent [term]
Corresponds to Metavar.

data FoilPattern (n :: S) (l :: S) where

Generated with mkFoilPattern. A scope-safe version of Pattern.

Constructors

FoilAPattern :: forall (n :: S) (l :: S). NameBinder n l -> FoilPattern n l
Corresponds to APattern.

Core operations

nfMetaTerm :: forall (n :: S) metavar. Distinct n => Scope n -> MetaTerm metavar n Type -> MetaTerm metavar n Type

nfMetaTermWithEmptyScope :: MetaTerm metavar 'VoidS Type -> MetaTerm metavar 'VoidS Type

matchPattern :: forall t (n :: S) (l :: S). InjectName t => FoilPattern n l -> t n -> Substitution t l n

Match a pattern against an term.

whnf :: forall (n :: S). Distinct n => Scope n -> Term n -> Term n

nf :: forall (n :: S). Distinct n => Scope n -> Term n -> Term n

Normalize a term to weak head normal form (WHNF). >>> nf Foil.emptyScope "λy:t.λz:f. (λx:t.λy:f.y) y z" λ x0 : t . λ x1 : f . x1 >>> nf Foil.emptyScope "λz:t.λw:u.(λx:t.λy:u->u.y) z (λz:u.z) w" λ x0 : t . λ x1 : u . x1 >>> nf Foil.emptyScope "(λb:t->f->t.λx:f.λy:t.b y x) (λx:t.λy:f.x)" λ x1 : f . λ x2 : t . x2 >>> nf Foil.emptyScope "(λs:(t->t)->t->t.λz:t->t.s(s z))(λb:t->t->t.λx:t.λy:t.b y x)(λx:t.λy:t.y)" λ x2 : t . λ x3 : t . x3 >>> nf Foil.emptyScope "(λs:(t->t)->t->t.λz:t->t.s (s z)) (λs:(t->t)->t->t.λz:t->t.s (s z)) (λb:t->t->t.λy:t.λx:t.b x y) (λy:t.λx:t.x)" λ x1 : t . λ x3 : t . x3 >>> nf Foil.emptyScope "(λ a : (t -> t) . λ x: u . x) (λ a : t . a)" λ x1 : u . x1

matchMetaAbs :: forall (binder :: S -> S -> Type) (sig :: Type -> Type -> Type) (ext :: Type -> Type -> Type). (UnifiablePattern binder, SinkableK binder, Bitraversable sig, Bitraversable ext, ZipMatchK sig, ZipMatchK ext, TypedBinder binder Type) => [Type] -> MetavarBinders -> MetaAbs (AnnBinder Type binder) (AnnSig Type (Sum sig (MetaAppSig MetavarIdent))) Type -> MetaAbs (AnnBinder Type binder) (AnnSig Type (Sum sig ext)) Type -> [MetaSubsts (AnnBinder Type binder) (AnnSig Type (Sum sig ext)) MetavarIdent Type]

moreGeneralThan :: MetavarBinders -> MetaSubsts' -> MetaSubsts' -> Bool

Check if the left substitution is more general than the right substitution. Returns a list of booleans, one per metavariable in LHS substitution. A substitution σ1 is more general than σ2 if there exists a substitution τ such that σ2 = τ ∘ σ1 >>> rawMetavarBinders = ["M : [t] t -> t -> t", "N : [t] t -> t"] >>> Right metavarBinders = parseMetavarBinders rawMetavarBinders >>> Right lhsSubst = parseMetaSubst metavarBinders "M[x] ↦ λy:t.N[x]" >>> Right rhsSubst = parseMetaSubst metavarBinders "M[x] ↦ λy:t.λz:t.x" >>> lhsSubsts = MetaSubsts [lhsSubst] >>> rhsSubsts = MetaSubsts [rhsSubst] >>> moreGeneralThan metavarBinders lhsSubsts rhsSubsts True >>> rawMetavarBinders = ["M : [t] t -> t -> t", "N : [t] t -> t"] >>> Right metavarBinders = parseMetavarBinders rawMetavarBinders >>> Right lhsSubst = parseMetaSubst metavarBinders "M[x] ↦ λy:t.λz:t.x" >>> lhsSubsts = MetaSubsts [lhsSubst] >>> rhsSubsts = MetaSubsts [] >>> moreGeneralThan metavarBinders lhsSubsts rhsSubsts False >>> rawMetavarBinders = ["F: [t -> t, t] t", "X: [t -> t] t", "Y: [] t", "H: [t -> t, t] t"] >>> Right metavarBinders = parseMetavarBinders rawMetavarBinders >>> Right lhsSubst = parseMetaSubst metavarBinders "M[x] ↦ λy:t.λz:t.x" >>> lhsSubsts = MetaSubsts [lhsSubst] >>> rhsSubsts = MetaSubsts [] >>> moreGeneralThan metavarBinders lhsSubsts rhsSubsts interactive:1:2-68: Non-exhaustive patterns in Right lhsSubst

toBinders :: forall (i :: S) (n :: S). Distinct i => NameBinderList i n -> NameMap n Type -> [VarBinder]

withMetaSubstVars :: forall (n :: S) (i :: S) r. Distinct n => Scope n -> Map VarIdent (Name n) -> NameBinderList i n -> NameMap n Type -> [(VarIdent, Type)] -> (forall (l :: S). Distinct l => Scope l -> Map VarIdent (Name l) -> NameBinderList i l -> NameMap l Type -> r) -> r

fromTerm :: forall (n :: S). Term n -> Term

fromMetaTerm :: forall (n :: S) t. MetaTerm MetavarIdent n t -> Term n

fromMetaSubst :: MetaSubst' -> MetaSubst

toMetaSubst :: MetavarBinders -> MetaSubst -> Maybe MetaSubst'

annotate :: forall (n :: S). MetavarBinders -> NameMap n Type -> Term n -> Maybe (MetaTerm MetavarIdent n Type, Type)

toTerm :: forall (n :: S). Distinct n => Scope n -> Map VarIdent (Name n) -> Term -> Term n

Convert Term' into a scope-safe term. This is a special case of convertToAST.

toTermClosed :: Term -> Term 'VoidS

Convert Term' into a closed scope-safe term. This is a special case of toTerm'.

getTermFromScopedTerm :: ScopedTerm -> Term

Generated with mkGetScopedTerm. Extract scoped term or crash.

getPatternBinder :: Pattern -> Maybe VarIdent

Generated with mkGetPatternBinder. Extract at most one binder from a pattern or crash.

toIdents :: forall (i :: S) (n :: S). Distinct i => NameBinderList i n -> [VarIdent]

Pattern synonyms

pattern App' :: forall binder (q :: Type -> Type -> Type) n. AST binder (AnnSig Type (Sum TermSig q)) n -> AST binder (AnnSig Type (Sum TermSig q)) n -> Type -> AST binder (AnnSig Type (Sum TermSig q)) n

pattern Lam' :: forall binder n (q :: Type -> Type -> Type) l. () => binder n l -> Type -> AST binder (AnnSig Type (Sum TermSig q)) l -> Type -> AST binder (AnnSig Type (Sum TermSig q)) n

pattern MetaVar' :: forall binder (q :: Type -> Type -> Type) n. MetavarIdent -> [AST binder (AnnSig Type (Sum TermSig q)) n] -> Type -> AST binder (AnnSig Type (Sum TermSig q)) n

pattern MetaApp :: forall metavar (binder :: S -> S -> Type) (sig :: Type -> Type -> Type) n t. metavar -> [TypedSOAS binder metavar sig n t] -> t -> TypedSOAS binder metavar sig n t

A convenient pattern synonym for parametrised metavariables.

Re-exports

alphaEquiv :: forall (sig :: Type -> Type -> Type) (n :: S) (binder :: S -> S -> Type). (Bitraversable sig, ZipMatchK sig, Distinct n, UnifiablePattern binder, SinkableK binder) => Scope n -> AST binder sig n -> AST binder sig n -> Bool

\(\alpha\)-equivalence check for two terms in one scope via unification of bound variables (via unifyNameBinders).

Compared to alphaEquivRefreshed, this function might skip unnecessary changes of bound variables when both binders in two matching scoped terms coincide.

Orphan instances

IsString MetavarBinder

Instance details

Methods

fromString :: String -> MetavarBinder

ZipMatchK MetavarIdent

Instance details

Methods

zipMatchWithK :: forall (as :: LoT Type) (bs :: LoT Type) (cs :: LoT Type). Mappings as bs cs -> (MetavarIdent :@@: as) -> (MetavarIdent :@@: bs) -> Maybe (MetavarIdent :@@: cs)

ZipMatchK Type

Instance details

Methods

zipMatchWithK :: forall (as :: LoT Type) (bs :: LoT Type) (cs :: LoT Type). Mappings as bs cs -> (Type :@@: as) -> (Type :@@: bs) -> Maybe (Type :@@: cs)

ZipMatchK VarIdent
Match Ident via Eq.

Instance details

Methods

zipMatchWithK :: forall (as :: LoT Type) (bs :: LoT Type) (cs :: LoT Type). Mappings as bs cs -> (VarIdent :@@: as) -> (VarIdent :@@: bs) -> Maybe (VarIdent :@@: cs)