Data.SOAS

Documentation

data MetaAppSig metavar scope term

Second-order signature for metavariable applications. This way, metavariables can be added to any other signature using Sum.

Constructors

MetaAppSig metavar [term]
A metavariable is always fully applied to a list of arguments.

type SOAS (binder :: S -> S -> Type) metavar (sig :: Type -> Type -> Type) (n :: S) = AST binder (Sum sig (MetaAppSig metavar)) n

A second-order abstract syntax (SOAS) is generated from a signature by adding parametrised metavariables.

Note that here we also parametrise SOAS by the type of binders (patterns) binder.

type ScopedSOAS (binder :: S -> S -> Type) metavar (sig :: Type -> Type -> Type) (n :: S) = ScopedAST binder (Sum sig (MetaAppSig metavar)) n

A scoped version of SOAS. ScopedSOAS is to SOAS what ScopedAST is to AST.

pattern MetaApp :: forall metavar binder (sig :: Type -> Type -> Type) n. metavar -> [SOAS binder metavar sig n] -> SOAS binder metavar sig n

A convenient pattern synonym for parametrised metavariables.

data MetaAbs (binder :: S -> S -> Type) (sig :: Type -> Type -> Type) t where

A body of a metavariable substitution for one metavariable.

Constructors

MetaAbs

Fields

:: forall (n :: S) t (binder :: S -> S -> Type) (sig :: Type -> Type -> Type). NameBinderList 'VoidS n

A list of binders corresponding to metavariable arguments.

-> NameMap n t

Types of parameters.

-> AST binder sig n

Term to substitute the metavariable with.

-> MetaAbs binder sig t

newtype MetaSubst (binder :: S -> S -> Type) (sig :: Type -> Type -> Type) metavar (ext :: Type -> Type -> Type) t

A metavariable substitution is a pair of a metavariable name and its body.

Constructors

MetaSubst

Fields

getMetaSubst :: (metavar, MetaAbs binder (Sum sig ext) t)

newtype MetaSubsts (binder :: S -> S -> Type) (sig :: Type -> Type -> Type) metavar (ext :: Type -> Type -> Type) t

A collection of metavariable substitutions (for simultaneous substitution of multiple metavariables).

Constructors

MetaSubsts

Fields

getMetaSubsts :: [MetaSubst binder sig metavar ext t]

applyMetaSubsts

Arguments

:: forall (sig :: Type -> Type -> Type) metavar (n :: S) (binder :: S -> S -> Type) metavar' t. (Bifunctor sig, Eq metavar, Distinct n, CoSinkable binder, SinkableK binder)

=> (metavar -> metavar')
Renaming of metavariables (normally, id or some kind of injection).

-> Scope n
Scope of terms.

-> MetaSubsts binder sig metavar (MetaAppSig metavar') t
Metavariable substitutions.

-> SOAS binder metavar sig n
The original SOAS term.

-> SOAS binder metavar' sig n

Apply metavariable substitutions to a SOAS term.

combineMetaSubsts :: forall metavar (sig :: Type -> Type -> Type) (ext :: Type -> Type -> Type) (binder :: S -> S -> Type) t. (Eq metavar, Bitraversable sig, Bitraversable ext, ZipMatchK (Sum sig ext), UnifiablePattern binder, SinkableK binder) => [MetaSubsts binder sig metavar ext t] -> [MetaSubsts binder sig metavar ext t]

Combine (compose) metavariable substitutions.

TODO: refactor

match

Arguments

:: forall (sig :: Type -> Type -> Type) metavar (n :: S) (binder :: S -> S -> Type) (ext :: Type -> Type -> Type) t. (Bitraversable sig, ZipMatchK sig, Eq metavar, Distinct n, UnifiablePattern binder, SinkableK binder, Bitraversable ext, ZipMatchK (Sum sig ext))

=> Scope n
The current scope.

-> SOAS binder metavar sig n
The left-hand side (with metavariables that we solve for).

-> AST binder (Sum sig ext) n
The right hand side (rigid).

-> [MetaSubsts binder sig metavar ext t]

Match left-hand side (with metavariables) against the rigid right-hand side.

If matching is successful, it produces metavariable substitutions that when applied to LHS make it syntactically equal to RHS. For example, matching M[f x, g] = g (f x) produces substitution M[z₁, z₂] ↦ z₂ z₁

There may be more than one solution for matching, e.g. M[f x, f x] = f x can be solved with two different substitutions: 1. M[z₁, z₂] ↦ z₁ 2. M[z₁, z₂] ↦ z₂

Hence, this function produces a list of possible substitutions.

matchScoped :: forall (sig :: Type -> Type -> Type) (n :: S) metavar (binder :: S -> S -> Type) (ext :: Type -> Type -> Type) t. (Bitraversable sig, ZipMatchK sig, Distinct n, Eq metavar, UnifiablePattern binder, SinkableK binder, Bitraversable ext, ZipMatchK (Sum sig ext)) => Scope n -> ScopedSOAS binder metavar sig n -> ScopedAST binder (Sum sig ext) n -> [MetaSubsts binder sig metavar ext t]

Same as match but for scoped terms.

matchMetavar

Arguments

:: forall metavar (n :: S) (m :: S) (binder :: S -> S -> Type) (sig :: Type -> Type -> Type) (ext :: Type -> Type -> Type) t. (Eq metavar, Distinct n, Distinct m, UnifiablePattern binder, SinkableK binder, Bitraversable sig, Bitraversable ext, ZipMatchK sig, ZipMatchK (Sum sig ext))

=> Scope m
Scope for the body of the metavariable substitution (at root of LHS).

-> NameBinderList 'VoidS m
List of binders for the metavariable parameters (size of the list should match the actual list of parameters).

-> Scope n
Current scope.

-> [SOAS binder metavar sig n]
A list of arguments of the parametrised metavariable on the left-hand side.

-> AST binder (Sum sig ext) n
The right-hand side term (rigid).

-> [(AST binder (Sum sig ext) m, MetaSubsts binder sig metavar ext t)]

A special case of match, when LHS is a parametrised metavariable.

Note that here we do not have an explicit name for that metavariable, and only consider its arguments. This is helpful when recursively matching with fresh metavariables, since we do not need to generate any actual fresh names for such metavariables, saving in complexity and performance.

For each possible solution, this function produces a pair of

Body of the metavariable substitution.

Metavariable substitutions for the metavariables that occur in the parameters of the root metavariable on the LHS.

matchMetavarScoped :: forall (n :: S) (m :: S) metavar (binder :: S -> S -> Type) (sig :: Type -> Type -> Type) (ext :: Type -> Type -> Type) t. (Distinct n, Distinct m, Eq metavar, UnifiablePattern binder, Bitraversable sig, Bitraversable ext, ZipMatchK sig, ZipMatchK (Sum sig ext), SinkableK binder) => Scope m -> NameBinderList 'VoidS m -> Scope n -> [SOAS binder metavar sig n] -> ScopedAST binder (Sum sig ext) n -> [(ScopedAST binder (Sum sig ext) m, MetaSubsts binder sig metavar ext t)]

Same as matchMetavar but for scoped term.

withFreshNameBinderList :: forall (n :: S) a (i :: S) r. Distinct n => [a] -> Scope n -> NameBinderList i n -> (forall (l :: S). Distinct l => Scope l -> NameBinderList i l -> r) -> r

Generate fresh name binders for a list of things. This is useful to generate proper name binders of metavariable parameters.

push :: forall (i :: S) (l :: S) (n :: S). NameBinder i l -> NameBinderList n i -> NameBinderList n l

O(n). Push a name binder into the end of a name binder list.

Should be in Control.Monad.Foil.

concatNameBinderLists :: forall (i :: S) (l :: S) (n :: S). NameBinderList i l -> NameBinderList n i -> NameBinderList n l

O(n). Concatenate name binder lists.

Should be in Control.Monad.Foil.