Language.Lambda.Impl

Description

Free foil implementation of the (lambda)-calculus (with pairs).

Free foil provides general definitions or implementations for the following:

Freely generated (from a simple signature) scope-safe AST.

Correct capture-avoiding substitution (see substitute).

Correct α-equivalence checks (see alphaEquiv and alphaEquivRefreshed) as well as α-normalization (see refreshAST).

Conversion helpers (see convertToAST and convertFromAST).

The following is generated using Template Haskell:

Convenient pattern synonyms.

ZipMatch instances (enabling general α-equivalence).

Conversion between scope-safe and raw term representation.

The following is implemented manually in this module:

Computation of weak head normal form (WHNF), see whnf.

Entry point, gluing everything together. See defaultMain.

Note: free foil does not (easily) support patterns at the moment, so only wildcard patterns and variable patterns are handled in this implementation.

Documentation

>>> :set -XOverloadedStrings
>>> :set -XDataKinds
>>> import qualified Control.Monad.Foil as Foil
>>> import Control.Monad.Free.Foil
>>> import Data.String (fromString)

Generated code

Signature

data TermSig scope term

Generated with mkSignature. A signature bifunctor, specifying the nodes of a syntax tree corresponding to Term.

Constructors

LamSig Type scope
Corresponds to Lam.

LetSig term scope
Corresponds to Let.

AppSig term term
Corresponds to App.

MetaVarSig MetaVarIdent [term]
Corresponds to MetaVar.

Pattern synonyms

pattern MetaVar :: MetaVarIdent -> [AST binder TermSig n] -> AST binder TermSig n

Generated with mkPatternSynonyms. Pattern synonym for an AST node of type MetaVarSig.

pattern App :: AST binder TermSig n -> AST binder TermSig n -> AST binder TermSig n

Generated with mkPatternSynonyms. Pattern synonym for an AST node of type AppSig.

pattern Let :: () => AST binder TermSig n -> binder n l1 -> AST binder TermSig l1 -> AST binder TermSig n

Generated with mkPatternSynonyms. Pattern synonym for an AST node of type LetSig.

pattern Lam :: () => Type -> binder n l1 -> AST binder TermSig l1 -> AST binder TermSig n

Generated with mkPatternSynonyms. Pattern synonym for an AST node of type LamSig.

Conversion helpers

convertToTermSig :: Term -> Either VarIdent (TermSig (Pattern, ScopedTerm) Term)

Generated with mkConvertToSig. Perform one step of converting Term, peeling off one node of type TermSig.

getPatternBinder :: Pattern -> Maybe VarIdent

Generated with mkGetPatternBinder. Extract at most one binder from a pattern or crash.

getTermFromScopedTerm :: ScopedTerm -> Term

Generated with mkGetScopedTerm. Extract scoped term or crash.

convertFromTermSig :: TermSig (Pattern, ScopedTerm) Term -> Term

Generated with mkConvertFromSig. Perform one step of converting Term, peeling off one node of type TermSig.

Scope-safe patterns

data FoilPattern (n :: S) (l :: S) where

Generated with mkFoilPattern. A scope-safe version of Pattern.

Constructors

FoilAPattern :: forall (n :: S) (l :: S). NameBinder n l -> FoilPattern n l
Corresponds to APattern.

toFoilPattern :: forall (n :: S) r. Distinct n => Scope n -> Map VarIdent (Name n) -> Pattern -> (forall (l :: S). DExt n l => FoilPattern n l -> Map VarIdent (Name l) -> r) -> r

Generated with mkToFoilPattern. Convert a raw pattern into a scope-safe pattern.

fromFoilPattern :: forall (n :: S) (l :: S). (Int -> VarIdent) -> FoilPattern n l -> Pattern

Generated with mkFromFoilPattern. Convert a scope-safe pattern into a raw pattern.

Pattern synonyms

pattern App' :: forall binder (q :: Type -> Type -> Type) n. AST binder (Sum TermSig q) n -> AST binder (Sum TermSig q) n -> AST binder (Sum TermSig q) n

pattern Lam' :: forall binder n (q :: Type -> Type -> Type) l. () => Type -> binder n l -> AST binder (Sum TermSig q) l -> AST binder (Sum TermSig q) n

pattern Let' :: forall binder (q :: Type -> Type -> Type) n l. () => AST binder (Sum TermSig q) n -> binder n l -> AST binder (Sum TermSig q) l -> AST binder (Sum TermSig q) n

pattern MetaVar' :: forall binder (q :: Type -> Type -> Type) n. MetaVarIdent -> [AST binder (Sum TermSig q) n] -> AST binder (Sum TermSig q) n

User-defined code

type AST' = AST FoilPattern

type Term = AST' TermSig

Scope-safe λ-term representation in scope n.

type MetaTerm metavar (n :: S) = SOAS FoilPattern metavar TermSig n

nfMetaTerm :: forall (n :: S) metavar. Distinct n => Scope n -> MetaTerm metavar n -> MetaTerm metavar n

nfMetaTermWithEmptyScope :: SOAS FoilPattern metavar TermSig 'VoidS -> SOAS FoilPattern metavar TermSig 'VoidS

nameMapToSubsts :: forall (i :: S) e (o :: S). NameMap i (e o) -> Substitution e i o

Conversion helpers for MetaSubst

toMetaSubst :: MetaSubst -> MetaSubst'

withMetaSubstVars :: forall (n :: S) (i :: S) r. Distinct n => [Binder] -> Scope n -> Map VarIdent (Name n) -> NameBinderList i n -> NameMap n Type -> (forall (l :: S). Distinct l => Scope l -> Map VarIdent (Name l) -> NameBinderList i l -> NameMap l Type -> r) -> r

type MetaSubst' = MetaSubst FoilPattern TermSig MetaVarIdent (MetaAppSig MetaVarIdent) Type

type MetaSubsts' = MetaSubsts FoilPattern TermSig MetaVarIdent (MetaAppSig MetaVarIdent) Type

fromMetaSubst :: MetaSubst' -> MetaSubst

toBinders :: forall (i :: S) (n :: S). Distinct i => NameBinderList i n -> NameMap n Type -> [Binder]

data UnificationConstraint where

Constructors

UnificationConstraint :: forall (n :: S). Distinct n => Scope n -> NameBinderList 'VoidS n -> NameMap n Type -> MetaTerm MetaVarIdent n -> MetaTerm MetaVarIdent n -> UnificationConstraint

toUnificationConstraint :: UnificationConstraint -> UnificationConstraint

fromUnificationConstraint :: UnificationConstraint -> UnificationConstraint

Conversion helpers for MetaTerm

toMetaTerm :: forall (n :: S). Term n -> MetaTerm MetaVarIdent n

fromMetaTerm :: forall (n :: S). MetaTerm MetaVarIdent n -> Term n

Conversion helpers

toTerm :: forall (n :: S). Distinct n => Scope n -> Map VarIdent (Name n) -> Term -> Term n

Convert Term' into a scope-safe term. This is a special case of convertToAST.

toTermClosed :: Term -> Term 'VoidS

Convert Term' into a closed scope-safe term. This is a special case of toTerm'.

fromTerm :: forall (n :: S). Term n -> Term

lam :: forall (n :: S). Distinct n => Type -> Scope n -> (forall (l :: S). DExt n l => Name l -> Term l) -> Term n

lam' :: forall (n :: S) metavar. Distinct n => Type -> Scope n -> (forall (l :: S). DExt n l => Name l -> MetaTerm metavar l) -> MetaTerm metavar n

unsafeParseTerm :: String -> Term 'VoidS

parseMetaSubst :: String -> Either String MetaSubst'

unsafeParseMetaSubst :: String -> MetaSubst'

unsafeParseUnificationConstraint :: String -> UnificationConstraint

matchPattern :: forall t (n :: S) (l :: S). InjectName t => FoilPattern n l -> t n -> Substitution t l n

Match a pattern against an term.

whnf :: forall (n :: S). Distinct n => Scope n -> Term n -> Term n

nf :: forall (n :: S). Distinct n => Scope n -> Term n -> Term n

interpretCommand :: Command -> IO ()

interpretProgram :: Program -> IO ()

Test framework implementation

solveUnificationConstraint :: MetaSubsts FoilPattern TermSig MetaVarIdent (MetaAppSig MetaVarIdent) Type -> UnificationConstraint -> UnificationConstraint

isSolved :: UnificationConstraint -> Bool

data Config

Constructors

Config

Fields

configLanguage :: Text

configFragment :: Text

configProblems :: [Problem]

data Problem

Constructors

Problem

Fields

problemConstraints :: [UnificationConstraint]

problemSolutions :: [Solution]

data Solution

Constructors

Solution

Fields

solutionName :: Text

solutionSubstitutions :: [MetaSubst']

configCodec :: TomlCodec Config

problemCodec :: TomlCodec Problem

parseTextToEither :: (String -> Either String a) -> Text -> Either Text a

solutionCodec :: TomlCodec Solution

toCodec :: Show a => (String -> Either String a) -> Key -> TomlCodec a

validateProblem :: Problem -> ([(Solution, [UnificationConstraint])], [Solution])

validateSolution :: [UnificationConstraint] -> Solution -> Either (Solution, [UnificationConstraint]) Solution

printInvalidSolutionsWithConstraint :: (Foldable t, Show a) => (Solution, t a) -> IO ()

parseConfigAndValidate :: IO ()

main :: IO ()

Orphan instances

ZipMatchK MetaVarIdent

Instance details

Methods

zipMatchWithK :: forall (as :: LoT Type) (bs :: LoT Type) (cs :: LoT Type). Mappings as bs cs -> (MetaVarIdent :@@: as) -> (MetaVarIdent :@@: bs) -> Maybe (MetaVarIdent :@@: cs)

ZipMatchK Type

Instance details

Methods

zipMatchWithK :: forall (as :: LoT Type) (bs :: LoT Type) (cs :: LoT Type). Mappings as bs cs -> (Type :@@: as) -> (Type :@@: bs) -> Maybe (Type :@@: cs)

ZipMatchK VarIdent
Match Ident via Eq.

Instance details

Methods

zipMatchWithK :: forall (as :: LoT Type) (bs :: LoT Type) (cs :: LoT Type). Mappings as bs cs -> (VarIdent :@@: as) -> (VarIdent :@@: bs) -> Maybe (VarIdent :@@: cs)